fix some problems
