+++ onchat/src/onchat/StreamConnector.java
@@ -1,6 +1,5 @@
-package onchat;
+﻿package onchat;
 
-import java.net.*;
 import java.io.*;
 
 // StreamConnectorクラス
